requirements.txt
discord.py==2.3.2
matplotlib==3.8.2
pandas==2.1.4
python-dotenv==1.1.0